+++ the-mathmos-server/src/main/java/com/digirati/themathmos/service/impl/W3CAnnotationSearchServiceImpl.java
@@ -52,7 +52,7 @@ public class W3CAnnotationSearchServiceImpl extends AnnotationSearchServiceImpl 
 
 	List<W3CAnnotation> annotationList = annotationUtils.getW3CAnnotations(annoSearchArray);
 		
-	Map<String, Object> annoMap = annotationUtils.createAnnotationPage(queryString, annotationList, true, pagingParameters, this.getTotalHits(), false);
+	Map<String, Object> annoMap = annotationUtils.createAnnotationPage(queryString, annotationList, true, pagingParameters, (int)this.getTotalHits(), false);
 	   	
 	if(null != annoMap && !annoMap.isEmpty()){
